requirements.txt
langchain
langchain-openai
langchain-community
streamlit
python-dotenv
chromadb
tiktoken
pypdf
sentence-transformers
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
langchain-text-splitters
chromadb
faiss-cpu
langchain-groq
wikipedia
arxiv
langchainhub